+++ crates/core/src/lib.rs
@@ -32,6 +32,9 @@ pub use reth_stateless::{ExecutionWitness, StatelessTrie, UncompressedPublicKey}
 
 pub type EthEvmConfig<C> = reth_evm_ethereum::EthEvmConfig<C, EthEvmFactory>;
 
+#[cfg(feature = "cycle-tracker")]
+pub mod cycle_tracker;
+
 pub mod serde_bincode_compat {
     pub type Block<'a> = reth_primitives_traits::serde_bincode_compat::Block<
         'a,
@@ -73,6 +76,9 @@ where
 
     #[cfg(all(feature = "r0vm", target_os = "zkvm", target_vendor = "risc0"))]
     assert!(install_r0vm_crypto());
+
+    #[cfg(all(feature = "cycle-tracker", target_os = "zkvm", target_vendor = "risc0"))]
+    let config = cycle_tracker::guest::CycleTrackerEvmConfig::new(config);
 
     let (hash, _) = reth_stateless::stateless_validation_with_trie::<SparseState, _, _>(
         block, signers, witness, chain_spec, config,